moving DataBlock to separate file, adding tests
